fixed significant bugs (directory mtime)
du bug - include dir entry in calculation
approximate bug - was always rounding to 0 in the first decimal place
directory mtime bug - not all directories had the correct mtime

+++ src/statistics.cc
@@ -24,7 +24,7 @@ string oldMessage = ">24H old";
  */
 struct summaryStats {
     bool inProcess;
-    unsigned long lastBackupBytes;
+    size_t lastBackupBytes;
     time_t lastBackupTime;
     size_t totalUsed;
     size_t totalSaved;
